// src/app/types/movies.ts
export type Movie = {
  id: number
  backdrop_path: string
  genre_ids: number[]
  original_language: string
  original_title: string
  overview: string
  popularity: number
  poster_path: string
  release_date: string
  title: string
  vote_average: number
  vote_count: number
  revenue?: number
  runtime?: string
  status?: string
  genres?: Generes[]
}

export type MoviesDto = {
  page: number
  results: Movie[]
  total_pages: number
  total_results: number
}

export type Generes = {
  id: number
  name: string
}

export type GeneresDto = {
  genres : Generes[]
}